layers: Windows fixes to JSON gen
